RAG/requirements-rag.txt
openai>=1.10.0
supabase>=2.3.0
python-dotenv>=1.0.0
tiktoken>=0.5.2
pyyaml>=6.0.1
tqdm>=4.66.0
